"culture" attribute value added
To add to the inventory of attribute values, I plugged a "culture" one into the "value" type as a way to more concisely account for instances in which cultures or counter cultures are mentioned, without also directly mentioning another of the value types. For example, I used this to tag any tropes that deal with Mexican assimilation into American culture.