Attempted to add drop down button functionality
Hi all, I've attempted to add a drop down button functionality to our menu bar... Note in so doing I removed the unordered list and replaced them with <div class="dropdown">. Currently I have the methodology, home and analysis links as drop downs, which is a problem considering they are only one link. That we will need to fix.